move to utility
